Splitting up spectrum.py
Into several files, one for each (sub)class
